### 2023-02-02 #5121 @editorialbot
[PRE REVIEW]: ctbench: Compiler-assisted benchmarking for the study of C++ metaprogram compile times
**Submitting author:** @jpenuchot (<a href="http://orcid.org/0000-0002-6377-6880">Jules Pénuchot</a>)
**Repository:** https://github.com/JPenuchot/ctbench
**Branch with paper.md** (empty if default branch): joss
**Version:** 1.2.1
