### 2023-02-02 #5121 @danielskatz
@JPenuchot - thanks for your submission. Please add spaces before references in the .md, and work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.

Also, your paper is a bit longer than JOSS's normal 1000 word limit - perhaps there is content that could be put in the repo and pointed to, rather than being in the paper?
