### 2023-02-02 ef8c1c5
Added missing DOIs, fixed DOI with URL prefix
