+++ paper.bib
@@ -164,6 +164,7 @@
 @article{iglberger2012_1,
   author = {Klaus {Iglberger} and Georg {Hager} and Jan {Treibig} and Ulrich {R{
             \"u}de}},
+  doi = {10.1137/110830125},
   title = {Expression Templates Revisited: A Performance Analysis of Current
            Methodologies},
   journal = {SIAM Journal on Scientific Computing},
@@ -176,6 +177,7 @@
   author = {Klaus {Iglberger} and Georg {Hager} and Jan {Treibig} and Ulrich {R{
             \"u}de}},
   title = {High Performance Smart Expression Template Math Libraries},
+  doi = {10.1109/hpcsim.2012.6266939},
   booktitle = {Proceedings of the 2nd International Workshop on New Algorithms
                and Programming Models for the Manycore Era (APMM 2012) at HPCS
                2012},
@@ -192,7 +194,7 @@
   year = {2012},
   series = {Applications of GPU Computing Series},
   isbn = {978-0-12-385963-1},
-  doi = {https://doi.org/10.1016/B978-0-12-385963-1.00026-5},
+  doi = {10.1016/B978-0-12-385963-1.00026-5},
   url = {https://www.sciencedirect.com/science/article/pii/B9780123859631000265},
   author = {Nathan Bell and Jared Hoberock},
   abstract = {Publisher Summary This chapter demonstrates how to leverage the
