### 2023-02-02 9191118
Moved example from the paper to example/readme.md, added spacing at the end of the paper
