+++ example/readme.md
@@ -1,0 +1,159 @@
+# ctbench example
+
+This example covers a short yet practical example of ctbench usage. We want to
+calculate the sum of a series of integers known at compile-time, using a type
+template to store unsigned integer values at compile-time.
+
+We will be comparing the compile-time performance of two implementations:
+- one based on a recursive function template,
+- and one based on C++11 parameter pack expansion.
+
+First we need to include `utility` to instantiate our benchmark according to the
+size parameter using `std::make_index_sequence`, and define the compile-time
+container type for an unsigned integer:
+
+```cpp
+#include <utility>
+
+/// Compile-time std::size_t
+template <std::size_t N> struct ct_uint_t {
+  static constexpr std::size_t value = N;
+};
+```
+
+The first version of the metaprogram is based on a recursive template function:
+
+```cpp
+/// Recursive compile-time sum implementation
+template<typename ... Ts> constexpr auto sum();
+
+template <> constexpr auto sum() { return ct_uint_t<0>{}; }
+template <typename T> constexpr auto sum(T const &) { return T{}; }
+
+template <typename T, typename... Ts>
+constexpr auto sum(T const &, Ts const &...tl) {
+  return ct_uint_t<T::value + decltype(sum(tl...))::value>{};
+}
+```
+
+And the other version relies on C++11 parameter pack expansion:
+
+```cpp
+/// Expansion compile-time sum implementation
+template<typename ... Ts> constexpr auto sum();
+
+template <> constexpr auto sum() { return ct_uint_t<0>{}; }
+
+template <typename... Ts> constexpr auto sum(Ts const &...) {
+  return ct_uint_t<(Ts::value + ... + 0)>{};
+}
+```
+
+Both versions share the same interface, and thus the same driver code as well.
+The driver code takes care of scaling the benchmark according to
+`BENCHMARK_SIZE`, which is defined by **ctbench** through the CMake API:
+
+```cpp
+// Driver code
+
+template <typename = void> constexpr auto foo() {
+  return []<std::size_t... Is>(std::index_sequence<Is...>) {
+    return sum(ct_uint_t<Is>{}...);
+  }
+  (std::make_index_sequence<BENCHMARK_SIZE>{});
+}
+
+[[maybe_unused]] constexpr std::size_t result =
+    decltype(foo())::value;
+```
+
+The CMake code needed to run the benchmarks is the following:
+
+```cmake
+ctbench_add_benchmark(
+  variadic_sum.expansion variadic_sum/expansion.cpp ${BENCHMARK_START}
+  ${BENCHMARK_STOP} ${BENCHMARK_STEP} ${BENCHMARK_ITERATIONS})
+
+ctbench_add_benchmark(
+  variadic_sum.recursive variadic_sum/recursive.cpp ${BENCHMARK_START}
+  ${BENCHMARK_STOP} ${BENCHMARK_STEP} ${BENCHMARK_ITERATIONS})
+```
+
+Then a graph target can be declared:
+
+```cmake
+ctbench_add_graph(variadic_sum-compare-graph compare-all.json
+                  variadic_sum.expansion variadic_sum.recursive)
+```
+
+with `compare-all.json` containing the following:
+
+```json
+{
+  "plotter": "compare_by",
+  "legend_title": "Timings",
+  "x_label": "Benchmark size factor",
+  "y_label": "Time (µs)",
+  "draw_average": true,
+  "demangle": false,
+  "draw_points": false,
+  "width": 800,
+  "height": 400,
+  "key_ptrs": ["/name", "/args/detail"],
+  "value_ptr": "/dur",
+  "plot_file_extensions": [".svg"]
+}
+```
+
+This configuration file uses the `compare_by` plotter to generate one plot for
[…]
