### 2023-02-02 815019b
Refreshed paper plan summary after removing example
