### 2023-02-02 7c17534
Chopped off redundant and unnecessary bits
