### 2023-02-06 #5121 @JPenuchot
Hello, I've been looking at the list again to try and find people who could be familiar with C++ metaprogramming. I'm trying to look for people who have C++ among their preferred language and feel comfortable reviewing HPC-related publications although this isn't an HPC library but mostly a tool that might be useful for HPC library developers.

I've found vyasr could be familiar enough with metaprogramming to review this submission.

I'm still struggling a bit to find people familiar with metaprogramming, if you have any suggestion regarding my searching methodology that would be much appreciated.
