### 2023-02-17 #5165 @editorialbot
[REVIEW]: ctbench: Compiler-assisted benchmarking for the study of C++ metaprogram compile times
**Submitting author:** @jpenuchot (<a href="http://orcid.org/0000-0002-6377-6880">Jules Pénuchot</a>)
**Repository:** https://github.com/JPenuchot/ctbench
**Branch with paper.md** (empty if default branch): main
**Version:** 1.3.4
