### 2023-03-04 JPenuchot/ctbench#20 @r-barnes
cmake too new with conflicting requirements.
I'm on Ubuntu 22.04, which is not that old, and finding that cmake 3.22.1 is too old to meet your CMakeLists.txt requirements. Can these be relaxed?

Furthermore, CMakeLists.txt requires 3.25 while CMakePresets.json requires 3.23. Can you make these consistent?

https://github.com/openjournals/joss-reviews/issues/5165
