### 2023-03-04 JPenuchot/ctbench#22 @r-barnes
Make prereq software more apparent
It'd be handy if you made a note in the readme that running
```
sudo apt install nlohmann-json3-dev
```
installs a necessary prerequisite. Perhaps there are others as well?

https://github.com/openjournals/joss-reviews/issues/5165
