### 2023-03-04 JPenuchot/ctbench#23 @r-barnes
Consider removing LLVM dependency
If I edit `CMakeLists.txt` to read:
```
cmake_minimum_required(VERSION 3.22)
```
and then run
```
cmake -DCMAKE_BUILD_TYPE=RelWithDebInfo -GNinja ..
ninja
```
I see
```
/z/ctbench/grapher/lib/grapher/plotters/debug.cpp:1:10: fatal error: llvm/Support/raw_ostream.h: No such file or directory
    1 | #include <llvm/Support/raw_ostream.h>
      |          ^~~~~~~~~~~~~~~~~~~~~~~~~~~~
/z/ctbench/grapher/grapher-utils.cpp:1:10: fatal error: llvm/Support/CommandLine.h: No such file or directory
    1 | #include <llvm/Support/CommandLine.h>
      |          ^~~~~~~~~~~~~~~~~~~~~~~~~~~~
```
Is it possible to remove this LLVM dependency in favour of using the C++ STL instead? If not, is it possible to make the LLVM linking mechanism more robust?

https://github.com/openjournals/joss-reviews/issues/5165
