### 2023-03-05 JPenuchot/ctbench#23 @JPenuchot
Hello,

> Is it possible to remove this LLVM dependency in favour of using the C++ STL instead?

The LLVM can't be dropped so easily as the CLI code relies heavily on LLVM's [CommandLine support library](https://llvm.org/docs/CommandLine.html), which doesn't have any equivalent in the STL. It is also required for symbol demangling in the `compare_by` plotter.

> If not, is it possible to make the LLVM linking mechanism more robust?

I'm not sure what you mean by more robust? The `find_package` command is called with the `REQUIRED` tag (here: https://github.com/JPenuchot/ctbench/blob/main/cmake/dependencies.cmake) and the `grapher` target is being linked against the export target exposed by LLVM (here: https://github.com/JPenuchot/ctbench/blob/main/grapher/cmake/grapher-target.cmake). I'm very surprised that CMake didn't catch any error and I'm suspecting that Ubuntu's way of splitting packages might be responsible for that.

It seems like most of your issues are due to the project not being built on Ubuntu as I use Arch for development, so I will likely set up an additional GitHub Action to check that ctbench compiles well on a recent Ubuntu system. I'll investigate that next week and come back to you, thanks for bringing that up.
