### 2023-03-06 JPenuchot/ctbench#22 @JPenuchot
Hi, there are other dependencies. Ubuntu is a bit difficult to deal with as it splits packages in dev and non-dev parts. Once again the Ubuntu CI I'll set up will be useful to find out the minimal set of required packages and add it to the docs.
