### 2023-03-06 JPenuchot/ctbench#20 @JPenuchot
Sadly the CMake 3.25 requirement can't be relaxed. Generator expressions are the only consistent way to set the `CXX_COMPILER_LAUNCHER` property properly that is essential for ctbench to function properly, and [CMake 3.25 is the first version to allow that](https://cmake.org/cmake/help/latest/release/3.25.html#properties).

I know this is far from being a satisfying solution, but as a temporary fix CMake 3.25 or newer can be installed locally using [Spack](https://spack.io/). Please let me know if you are willing to try this out. In the meantime I'll look for a better solution.
