### 2023-03-06 JPenuchot/ctbench#20 @r-barnes
I can probably get cmake 3.25 installed, but I'm not sure if that level of effort is commensurate with your expected audience. If so, you're fine. If not, then having me do it for the review isn't really helping making your software accessible.

Can you sort out the conflicting requirements, though? cmake 3.23 versus 3.25 in the same project seems odd.
