### 2023-03-06 JPenuchot/ctbench#20 @JPenuchot
>I can probably get cmake 3.25 installed, but I'm not sure if that level of effort is commensurate with your expected audience. If so, you're fine. If not, then having me do it for the review isn't really helping making your software accessible.

The next release of Ubuntu will come next month [with CMake 3.25](https://packages.ubuntu.com/lunar/cmake), so I will keep the CMake 3.25 requirement. The Ubuntu CI will target this release and I'll be working on the other issues in the meantime.

I tried several ways to handle it in the past, and using the generator expression in this context made the handling of the compiler launcher location much easier and more consistent.

>Can you sort out the conflicting requirements, though? cmake 3.23 versus 3.25 in the same project seems odd.

Sure! I already pushed that change in the `ubuntu-compatibility` branch which is where I'm staging changes related to your issues. I will close this issue when Ubuntu 23.04 gets released or when `ubuntu-compatibility` gets merged, but you can close it now if you feel like the issue is solved.
