### 2023-03-06 e4ee464
Added draft for dependencies part of the doc
