+++ readme.md
@@ -46,6 +46,18 @@ generated by the `compare_by` plotter*
 
 ## Using ctbench
 
+### Dependencies
+
+- nlohmann_json >= 3.9.1
+- Sciplot
+- LLVM
+- fmt
+
+- Ubuntu package list (incomplete): `nlohmann-json3-dev libfmt-dev`
+- ArchLinux (+ AUR) package list: `nlohmann-json fmt llvm clang` (+ `sciplot-git`)
+
+Installing
+
 ### Installing ctbench
 
 ```sh
