### 2023-03-08 4bd7a43
Split run-tests action in two separate actions for test runs on Arch and Ubuntu
