### 2023-03-08 52aaf60
Changed Arch package names to Ubuntu package names in apt install command
