### 2023-03-08 8375260
Added -flto to link options
