### 2023-03-08 5999b6b
Added package list for Ubuntu and install instructions for Sciplot
