+++ readme.md
@@ -46,17 +46,29 @@ generated by the `compare_by` plotter*
 
 ## Using ctbench
 
-### Dependencies
+### Build prerequisites
 
-- nlohmann_json >= 3.9.1
-- Sciplot
-- LLVM
-- fmt
+Ubuntu compatibility starts from Ubuntu 23.04. The CMake API files rely on a
+feature that was introduced in CMake 3.25 ([generator expressions for
+`<LANG>_COMPILER_LAUNCHER`](
+https://cmake.org/cmake/help/latest/release/3.25.html#properties)), which is not
+available on older versions of Ubuntu.
 
-- Ubuntu package list (incomplete): `nlohmann-json3-dev libfmt-dev`
-- ArchLinux (+ AUR) package list: `nlohmann-json fmt llvm clang` (+ `sciplot-git`)
+Here is the list of Ubuntu packages needed to build ctbench: `catch2 clang cmake
+git libboost-all-dev libbtbb-dev libclang-dev libfmt-dev libllvm15 libtbb-dev
+libtbb12 llvm llvm-dev ninja-build nlohmann-json3-dev`
 
-Installing
+And here is the list of ArchLinux packages: `git cmake ninja clang llvm
+llvm-libs nlohmann-json fmt boost boost-libs tbb catch2`
+
+The Sciplot library is required too. It can be installed on ArchLinux using the
+[`sciplot-git` AUR package](https://aur.archlinux.org/packages/sciplot-git)
+(NB: the non-git package isn't up-to-date).
+
+On Ubuntu, [you can install it system-wide using CMake](
+https://sciplot.github.io/installation/#installation-using-cmake). You may
+also use [vcpkg](https://vcpkg.io/en/getting-started.html) to install it
+locally. This method has not been tested yet, but instructions are on the way.
 
 ### Installing ctbench
 
