### 2023-03-08 JPenuchot/ctbench#22 @JPenuchot
Hello, the Ubuntu 23.04 CI was setup successfully, I've added a minimal list of required Ubuntu packages to the documentation. Feel free to close if this satisfies your expectations or give additional feedback if something is missing.
