### 2023-03-08 1846651
Removed core LLVM component from cmake-doc-extractor linking
