### 2023-03-09 ee1303e
Added missing packages for vcpkg bootstrapping
