+++ .github/actions/run-tests-archlinux/entrypoint.sh
@@ -10,6 +10,7 @@ cmake --preset release -G Ninja \
   -DCTBENCH_ENABLE_TESTS=ON \
   -DCTBENCH_ENABLE_TRACY=OFF \
   -DCTBENCH_ENABLE_DOCS=OFF \
-  -DCMAKE_TOOLCHAIN_FILE=vcpkg/scripts/buildsystems/vcpkg.cmake
+  -DCMAKE_TOOLCHAIN_FILE=vcpkg/scripts/buildsystems/vcpkg.cmake \
+  -DVCPKG_OVERLAY_PORTS=""
 
 cmake --build --preset release --target test-all
