### 2023-03-09 ed0f334
Try adding fmt as a vcpkg dependency to fix the missing export bug
