### 2023-03-09 b45b4a3
Propagated vcpkg use to Ubuntu CI
