### 2023-03-09 4b807b8
Added missing packages for vcpkg bootstrapping
