### 2023-03-09 3fe579a
Added instructions to pull Sciplot using vcpkg
