### 2023-03-09 JPenuchot/ctbench#22 @JPenuchot
I'm closing the issue as both apt and vcpkg package dependencies are now provided in the doc.
