### 2023-03-09 JPenuchot/ctbench#20 @JPenuchot
`ubuntu-compatibility` was merged, CMake 3.25 is the minimum requirement from now on as it can't really be worked around.
