### 2023-03-09 7d4f697
Reworked prerequisites part of the documentation
