+++ readme.md
@@ -48,27 +48,23 @@ generated by the `compare_by` plotter*
 
 ### Build prerequisites
 
-Ubuntu compatibility starts from Ubuntu 23.04. The CMake API files rely on a
-feature that was introduced in CMake 3.25 ([generator expressions for
-`<LANG>_COMPILER_LAUNCHER`](
-https://cmake.org/cmake/help/latest/release/3.25.html#properties)), which is not
-available on older versions of Ubuntu.
+ArchLinux and Ubuntu 23.04 are officially supported as tests are compiled and
+executed on both of these Linux distributions. Others including Fedora or any
+other Linux distro that provides CMake 3.25 or higher should be compatible.
 
-Here is the list of Ubuntu packages needed to build ctbench: `catch2 clang cmake
-git libboost-all-dev libclang-dev libfmt-dev libllvm15 libtbb-dev libtbb12 llvm
-llvm-dev ninja-build nlohmann-json3-dev`
+- Required ArchLinux packages: `boost boost-libs catch2 clang cmake curl fmt git
+llvm llvm-libs ninja nlohmann-json tar tbb unzip zip`
 
-And here is the list of ArchLinux packages: `git cmake ninja clang llvm
-llvm-libs nlohmann-json fmt boost boost-libs tbb catch2`
+- Required Ubuntu packages: `catch2 clang cmake curl git libboost-all-dev
+libclang-dev libfmt-dev libllvm15 libtbb-dev libtbb12 llvm llvm-dev ninja-build
+nlohmann-json3-dev pkg-config tar unzip zip`
 
 The Sciplot library is required too. It can be installed on ArchLinux using the
 [`sciplot-git` AUR package](https://aur.archlinux.org/packages/sciplot-git)
-(NB: the non-git package isn't up-to-date).
-
-On Ubuntu, [you can install it system-wide using CMake](
-https://sciplot.github.io/installation/#installation-using-cmake). You may
-also use [vcpkg](https://vcpkg.io/en/getting-started.html) to install it
-locally:
+(NB: the non-git package isn't up-to-date). Otherwise, [you can install it for
+your whole system using CMake](
+https://sciplot.github.io/installation/#installation-using-cmake) or locally
+using [vcpkg](https://vcpkg.io/en/getting-started.html):
 
 ```sh
 git clone https://github.com/Microsoft/vcpkg.git
@@ -79,8 +75,8 @@ cmake --preset release \
   -DCMAKE_TOOLCHAIN_FILE=vcpkg/scripts/buildsystems/vcpkg.cmake
 ```
 
-Note: The `fmt` vcpkg dependency is needed as vcpkg may break its CMake
-integration from the system package manager for whatever reason.
+Note: The `fmt` dependency is needed, as vcpkg breaks fmt's CMake integration if
+you have it already installed.
 
 ### Installing ctbench
 
