### 2023-03-17 c070a16
Rewrote debug plotter: outputs a JSON file, has an optional parameter to disable console output
