### 2023-03-17 61a741d
Fixed instance list being an array instead of a map
