### 2023-03-20 94f28bf
Refactoring: get_values was renamed to filtered_values_sums and has more explicit variable names
