### 2023-03-20 72e7406
Refactoring: get_values was renamed to filtered_values_sums and has more explicit variable names
