### 2023-03-20 a58d4b4
Adapted compare plotter to get_values name change
