### 2023-03-20 e806337
Adapted stack plotter to get_values name change
