### 2023-03-20 326a16e
Implemented event filtering for compare_by plotter
