### 2023-03-20 bed7e44
Updated compare_by documentation
