### 2023-04-03 9aad81c
Added basic math functions for average and median computation + tests
