### 2023-04-03 59fadb8
Making use of new average and median functions in compare_by plotter code, and added median line drawing option in the parameters
