### 2023-04-03 30fb093
fixed median drawing, simplified conditions on line drawings (sciplot does not draw lines if vectors are empty)
