### 2023-04-03 cb8399d
Statistical computations now have double values as results
