### 2023-04-03 b51c1a7
Switch to double values for everything on Y axis except point coordinates + added stddev error bars
