### 2023-04-04 8d0eaef
Fixed x coordinates being added twice, added config for error bars and disabled them by default
