### 2023-04-04 103748d
Shortened curve labels
