### 2023-04-04 cece650
Added ctbench_add_benchmark_for_range and ctbench_add_benchmark_for_size_list to provide more ways to specify benchmark sizes
