### 2023-04-04 759c837
Restablished previous ctbench_add_benchmark implementation, ctbench_add_benchmark_for_range being now implemented as syntactic sugar on top of it
