### 2023-04-04 8ff1788
console I/O converted to fmt + attempt to get exit code from spawned process (Linux only)
