### 2023-04-04 ffe2691
Implemented error status check and passthrough for ttw
