### 2023-04-06 a2e1307
Fixed nonexistent subtarget when using ctbench_add_benchmark
