+++ cmake/dependencies.cmake
@@ -4,6 +4,7 @@
 # duplicate find_package calls and expose all dependencies in a single file.
 
 find_package(nlohmann_json 3.9.1 REQUIRED)
+find_package(Boost REQUIRED)
 find_package(sciplot REQUIRED)
 find_package(LLVM REQUIRED CONFIG)
 find_package(fmt REQUIRED)
