### 2023-05-19 f1073f4
added example compilation to CI
