### 2023-05-19 04f430f
add install command to CI script
