### 2023-05-19 10746c2
adding container image to yet again work around github's painful bugs
