### 2023-05-19 6679e6f
entrypoint now exits on errors
