### 2023-05-19 89cfe3c
add -flto to link options
