+++ cmake/ctbench-compile-opts.cmake
@@ -15,4 +15,6 @@ target_compile_options(
             -flto
             -fPIC)
 
+target_link_options(ctbench-compile-opts INTERFACE -flto)
+
 target_compile_features(ctbench-compile-opts INTERFACE cxx_std_20)
