### 2023-05-19 f2ae8d7
removed redundant compile and link options
