### 2023-05-26 b29e28c
added benchmark execution instructions for the example
