### 2023-05-29 3c2a54b
refactored cmake for proper export of the grapher library
