### 2023-05-29 7809dfc
going back to a static library at least for now
