+++ grapher/lib/grapher/predicates.cpp
@@ -1,4 +1,3 @@
-#include <bits/iterator_concepts.h>
 #include <regex>
 #include <string>
 
