+++ grapher/cmake/grapher-target.cmake
@@ -10,10 +10,11 @@ target_include_directories(grapher
 
 llvm_map_components_to_libnames(llvm_libs support)
 
+target_link_libraries(grapher PRIVATE TBB::tbb)
+
 target_link_libraries(grapher
   PUBLIC
     ${llvm_libs}
-    TBB::tbb
     ctbench-compile-opts
     fmt::fmt
     nlohmann_json::nlohmann_json
