+++ grapher/include/grapher/utils/error.hpp
@@ -2,7 +2,6 @@
 
 #include <cstdint>
 #include <cstdlib>
-#include <experimental/source_location>
 #include <source_location>
 #include <string_view>
 
