+++ grapher/include/grapher/utils/json.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-#include <experimental/source_location>
 #include <optional>
+#include <source_location>
 #include <string>
 #include <type_traits>
 
