### 2023-05-31 60e3033
adding gnuplot to Ubuntu package list
