### 2023-06-21 #5165 @JPenuchot
I'm sorry I completely forgot about the missing contributing guidelines. I'm working on it right now.
