### 2023-06-21 #5165 @JPenuchot
@dirmeier I added contribution guidelines and a CoC for the project. Please let me know if you feel like they need to be revised, and thanks again for your time reviewing the project.
