+++ CONTRIBUTING.md
@@ -13,9 +13,8 @@ an issue, creating a PR, reviewing, and merging the PR.
 To get an overview of the project, read the [README](README.md). Here are some
 resources to help you get started with open source contributions:
 
-- [Finding ways to contribute to open source on
-GitHub](https://docs.github.com/en/get-started/exploring-projects-on-github/
-finding-ways-to-contribute-to-open-source-on-github) - [Set up
+- [Finding ways to contribute to open source on GitHub](
+https://docs.github.com/en/get-started/exploring-projects-on-github/finding-ways-to-contribute-to-open-source-on-github) - [Set up
 Git](https://docs.github.com/en/get-started/quickstart/set-up-git) - [GitHub
 flow](https://docs.github.com/en/get-started/quickstart/github-flow) -
 [Collaborating with pull
