### 2023-07-18 378738f
Switching to std::chrono for time measurement
