### 2023-07-18 f8cb401
switched to using boost process with argument list instead of a shell command
