+++ compiler-launcher/compiler-launcher.cpp
@@ -18,8 +18,6 @@
 #include <cstdlib>
 #include <filesystem>
 #include <fstream>
-#include <numeric>
-#include <sstream>
 #include <string>
 #include <string_view>
 
