### 2023-07-18 55e1575
added build result upload to CI
