### 2023-07-18 a26410c
refactoring + get_timetrace_file does not exit on non-zero status code but forwards the exit code by returning it instead
