+++ CHANGELOG.md
@@ -1,0 +1,10 @@
+# Changelog
+
+## 1.3.4
+
+- Added changelog
+
+### compiler-launcher
+
+- Added `-h/-help/help` flag to display help in CLI
+- Fixed compiler execution time measurement for measurements without time-trace
