### 2023-07-20 a0b551c
refactored compare_by code: cut the plot function into smaller ones to make the plot method less complex
