### 2023-07-20 8fc9c57
config reading refactoring
