+++ grapher/lib/grapher/plotters/compare_by.cpp
@@ -1,6 +1,7 @@
 #include <algorithm>
 #include <filesystem>
 #include <fstream>
+#include <numeric>
 #include <string>
 #include <vector>
 
@@ -8,7 +9,7 @@
 
 #include <boost/container/small_vector.hpp>
 
-#include <llvm/Demangle/Demangle.h>
+#include <boost/core/demangle.hpp>
 
 #include <sciplot/sciplot.hpp>
 
@@ -49,7 +50,73 @@ struct process_event_parameters_t {
   benchmark_instance_t const &instance;
 };
 
+/// Set of X and Y coordinate vectors for all curves and points of a graph.
+struct coordinate_vectors_t {
+  // X axis coordinates for all curves
+  std::vector<grapher::value_t> x_curve;
+
+  // Average curve Y coordinate vectors
+  std::vector<double> y_average_curve;
+  std::vector<double> y_delta_curve;
+  std::vector<double> y_median_curve;
+
+  // Point X and Y coordinate vectors
+  std::vector<grapher::value_t> x_points;
+  std::vector<grapher::value_t> y_points;
+};
+
+/// Parameter list for the generate_plot function
+/// extracted into a struct for readability
+struct generate_plot_parameters_t {
+  std::filesystem::path const &plot_output_folder;
+  bool draw_average;
+  bool average_error_bars;
+  bool draw_points;
+  bool draw_median;
+  bool demangle;
+};
+
+// =============================================================================
+// Forward declarations
+
 /// Generate a curve for a given time-trace event and stores it in output_map.
+inline void process_event(curve_aggregate_map_t &output_map,
+                          grapher::json_t const &event,
+                          process_event_parameters_t const &parameters);
+
+/// Scans event data at value_pointer and generates curves for each key
+/// generated from key_pointers. The curves are stored in a nested map
+/// structure which is far from optimal but we're limited by gnuplot's
+/// performance anyway.
+curve_aggregate_map_t
+get_bench_curves(benchmark_set_t const &input,
+                 std::vector<json_t::json_pointer> const &key_pointers,
+                 json_t::json_pointer const &value_pointer,
+                 std::vector<predicate_t> filters = {});
+
+/// Transforms a key into a string that's usable as a path.
+std::string to_string(key_t const &key, bool demangle = true);
+
+/// Draws the curves and points for a given benchmark.
+inline void draw_bench_curves(sciplot::Plot2D &plot,
+                              coordinate_vectors_t const &coord_vectors,
+                              std::string const &bench_name,
+                              generate_plot_parameters_t const &parameters);
+
+/// Reads plot generation parameters from the config
+generate_plot_parameters_t
+get_plotgen_parameters(grapher::json_t const &config,
+                       std::filesystem::path const &dest);
+
+/// Function to generate one plot.
+/// NB: This function must remain free of config reading logic.
+inline void generate_plot(
+    curve_aggregate_map_t::const_iterator::value_type aggregate_key_value,
+    generate_plot_parameters_t const &parameters);
+
+// =============================================================================
+// Function definitions
+
 inline void process_event(curve_aggregate_map_t &output_map,
                           grapher::json_t const &event,
                           process_event_parameters_t const &parameters) {
@@ -76,15 +143,11 @@ inline void process_event(curve_aggregate_map_t &output_map,
   }
 }
 
-/// Scans event data at value_pointer and generates curves for each key
-/// generated from key_pointers. The curves are stored in a nested map
-/// structure which is far from optimal but we're limited by gnuplot's
-/// performance anyway.
 curve_aggregate_map_t
 get_bench_curves(benchmark_set_t const &input,
[…]
