### 2023-07-25 d9dc85c
added plotter_config back to the generate_plot_parameters_t structure
