### 2023-08-23 #5165 @danielskatz
@JPenuchot - I'm suggesting a bunch of changes in https://github.com/JPenuchot/ctbench/pull/33 - please check these carefully to ensure I'm not changing your intent, and then either merge them or let me know what you think is incorrect.  Then we can proceed.
