### 2023-08-23 #5165 @danielskatz
@JPenuchot - I see that my changes were merged into the main branch of your repo, but the paper is being built from the JOSS branch. What's the difference between the paper.md/bib in main and joss, other than my changes? Should my changes just be moved in the joss branch? or should we be building from main?
