### 2023-08-23 #5165 @danielskatz
Congratulations to @JPenuchot (Jules Pénuchot) and co-author on your publication!!

JOSS depends completely on volunteers and wouldn't be able to work without you
